requirements.txt
Flask==3.0.3
bcrypt==4.1.2
PyJWT==2.8.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
pypdf==6.9.2
